add refactor get launchdetail

+++ app/src/main/java/com/x5bartsoft/spacex/model/response/launchdetail/Diameter.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Diameter(
     @SerializedName("feet")
-    val feet: Int,
+    val feet: Double,
     @SerializedName("meters")
     val meters: Double
 ) : Parcelable
